+++ src/memory.rs
@@ -12,7 +12,8 @@ use k8s_openapi::apimachinery::pkg::api::resource::Quantity;
 
 use crate::error::{Error, OperatorResult};
 use std::{
-    ops::{Add, Div, Mul, Sub},
+    fmt::Display,
+    ops::{Add, AddAssign, Div, Mul, Sub, SubAssign},
     str::FromStr,
 };
 
@@ -71,6 +72,21 @@ impl FromStr for BinaryMultiple {
                 value: q.to_string(),
             }),
         }
+    }
+}
+
+impl Display for BinaryMultiple {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let out = match self {
+            BinaryMultiple::Kibi => "Ki",
+            BinaryMultiple::Mebi => "Mi",
+            BinaryMultiple::Gibi => "Gi",
+            BinaryMultiple::Tebi => "Ti",
+            BinaryMultiple::Pebi => "Pi",
+            BinaryMultiple::Exbi => "Ei",
+        };
+
+        out.fmt(f)
     }
 }
 
@@ -295,6 +311,14 @@ impl Div<f32> for MemoryQuantity {
     }
 }
 
+impl Div<MemoryQuantity> for MemoryQuantity {
+    type Output = f32;
+
+    fn div(self, rhs: MemoryQuantity) -> Self::Output {
+        self.value / rhs.value
+    }
+}
+
 impl Sub<MemoryQuantity> for MemoryQuantity {
     type Output = MemoryQuantity;
 
@@ -306,6 +330,12 @@ impl Sub<MemoryQuantity> for MemoryQuantity {
     }
 }
 
+impl SubAssign<MemoryQuantity> for MemoryQuantity {
+    fn sub_assign(&mut self, rhs: MemoryQuantity) {
+        self.value -= rhs.value;
+    }
+}
+
 impl Add<MemoryQuantity> for MemoryQuantity {
     type Output = MemoryQuantity;
 
@@ -314,6 +344,12 @@ impl Add<MemoryQuantity> for MemoryQuantity {
             value: self.value + rhs.scale_to(self.unit).value,
             unit: self.unit,
         }
+    }
+}
+
+impl AddAssign<MemoryQuantity> for MemoryQuantity {
+    fn add_assign(&mut self, rhs: MemoryQuantity) {
+        self.value += rhs.value;
     }
 }
 
@@ -361,11 +397,24 @@ impl TryFrom<Quantity> for MemoryQuantity {
         Self::try_from(&quantity)
     }
 }
+
 impl TryFrom<&Quantity> for MemoryQuantity {
     type Error = Error;
 
     fn try_from(quantity: &Quantity) -> OperatorResult<Self> {
         quantity.0.parse()
+    }
+}
+
+impl From<MemoryQuantity> for Quantity {
+    fn from(quantity: MemoryQuantity) -> Self {
+        Self::from(&quantity)
+    }
+}
+
+impl From<&MemoryQuantity> for Quantity {
+    fn from(quantity: &MemoryQuantity) -> Self {
+        Quantity(format!("{}{}", quantity.value, quantity.unit))
     }
 }